Updating json, modify models, modify html

+++ mywatchlist/models.py
@@ -4,6 +4,6 @@ from django.db import models
 class MyWatchList(models.Model):
     watched = models.BooleanField(default=False)
     title = models.CharField(max_length=255)
-    rating = models.IntegerField()
+    rating = models.FloatField()
     release_date = models.DateField()
     review = models.TextField()
